build: moving static cluster into extensions directory
