+++ source/extensions/clusters/static/static_cluster.cc
@@ -1,4 +1,4 @@
-#include "source/common/upstream/static_cluster.h"
+#include "source/extensions/clusters/static/static_cluster.h"
 
 #include "envoy/common/exception.h"
 #include "envoy/config/cluster/v3/cluster.pb.h"
