+++ test/server/BUILD
@@ -54,6 +54,7 @@ envoy_cc_test(
         "//source/common/http:context_lib",
         "//source/common/upstream:cluster_manager_lib",
         "//source/extensions/access_loggers/file:config",
+        "//source/extensions/clusters/static:static_cluster_lib",
         "//source/extensions/stat_sinks/statsd:config",
         "//source/extensions/transport_sockets/raw_buffer:config",
         "//source/server:configuration_lib",
