+++ tools/code_format/config.yaml
@@ -291,6 +291,7 @@ unsorted_flags:
 # PLEASE DO NOT ADD FILES TO THIS LIST WITHOUT SENIOR MAINTAINER APPROVAL
 visibility_excludes:
 - source/extensions/clusters/strict_dns/
+- source/extensions/clusters/static/
 - source/extensions/clusters/original_dst/
 - source/extensions/clusters/logical_dns/
 - source/extensions/early_data/BUILD
